requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
faiss-cpu==1.8.0.post1
fastapi==0.114.2
greenlet==3.1.0
h11==0.14.0
idna==3.9
numpy==1.26.4
packaging==24.1
pydantic==2.9.1
pydantic_core==2.23.3
redis==5.0.8
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
